Rename src/language-server/code_completion.ts -> src/code_completion.ts
